Move new file search code to own directory
