+++ app/src/main/java/net/gsantner/markor/ui/fsearch/FileSearchDialog.java
@@ -1,4 +1,4 @@
-package net.gsantner.opoc.ui;
+package net.gsantner.markor.ui.fsearch;
 
 import android.app.Activity;
 import android.support.v4.content.ContextCompat;
